requirements-dev.txt
# Development dependencies for YARA AST
# Install with: pip install -r requirements-dev.txt

# Core dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.0.0
hypothesis>=6.0.0

# Code formatting and linting
black>=23.0.0
isort>=5.12.0
flake8>=7.0.0
flake8-docstrings>=1.7.0
flake8-bugbear>=23.0.0
flake8-comprehensions>=3.14.0
flake8-simplify>=0.20.0
ruff>=0.1.0

# Type checking
mypy>=1.8.0
types-PyYAML>=6.0.0
types-requests>=2.31.0

# Security and code quality
bandit>=1.7.0
pydocstyle>=6.3.0
safety>=2.3.0

# Pre-commit hooks
pre-commit>=3.0.0

# Quality tools
qlty>=0.1.0

# All optional features for development
yara-python>=4.5.0
psutil>=5.9.0
graphviz>=0.20.0
